chore: Add BackButton component to Header
Added the BackButton component to the Header component. The BackButton component is conditionally rendered based on the showBackButton prop. It includes a BackIcon component, which utilizes the CaretLeft icon from the phosphor-react-native library. Styled the BackButton component as a TouchableOpacity with flex 1. Set the size and color of the BackIcon using styled-components attrs method.